fix: renomear arquivo e ajustar nas rotas
